Phase 3/requirements.txt
numpy
pandas
PyYAML
matplotlib
imageio
imageio-ffmpeg
mujoco
jupyterlab
ipykernel
